+++ src/Exceptions/DuplicateColumnKeyException.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 /**
  * Text Table
  *
- * @package   MarcinOrlowski\TextTable
- *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/php-text-table
  */
@@ -16,5 +13,9 @@ namespace MarcinOrlowski\TextTable\Exceptions;
 
 class DuplicateColumnKeyException extends \Exception
 {
-
+    public static function forColumnKey(string|int $columnKey): static
+    {
+        $msg = \sprintf('Duplicate column key: %s', $columnKey);
+        return self($msg);
+    }
 }