+++ src/Exceptions/UnsupportedColumnTypeException.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 /**
  * Text Table
  *
- * @package   MarcinOrlowski\TextTable
- *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/php-text-table
  */
@@ -16,5 +13,10 @@ namespace MarcinOrlowski\TextTable\Exceptions;
 
 class UnsupportedColumnTypeException extends \Exception
 {
-
+    public static function forColumnKeyVal(string $columnKey, mixed $columnVal): static
+    {
+        $msg = \sprintf('Unsupported column type (%s): %s', \get_debug_type($columnVal), $columnKey);
+        return self($msg);
+    }
+// empty
 }