+++ src/Renderers/RendererContract.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 /**
  * Text Table
  *
- * @package   MarcinOrlowski\TextTable
- *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/php-text-table
  */
@@ -29,8 +26,6 @@ interface RendererContract
      * Helper that returns rendered table as single string.
      *
      * @param TextTable $table Instance of `TextTable` to render.
-     *
-     * @return string
      */
     public function renderAsString(TextTable $table): string;
 }