+++ src/Row.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 /**
  * Text Table
  *
- * @package   MarcinOrlowski\TextTable
- *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/php-text-table
  */
@@ -17,19 +14,14 @@ namespace MarcinOrlowski\TextTable;
 use Lombok\Getter;
 use Lombok\Setter;
 use MarcinOrlowski\TextTable\Exceptions\DuplicateColumnKeyException;
-use MarcinOrlowski\TextTable\Traits\ArrayAccessTrait;
-use MarcinOrlowski\TextTable\Traits\IteratorAggregateTrait;
 
 /**
  * @method CellsContainer getContainer()
  */
 #[Getter]
 #[Setter]
-class Row extends \Lombok\Helper implements ContainerContract
+class Row extends BaseContainer
 {
-    use ArrayAccessTrait;
-    use IteratorAggregateTrait;
-
     /**
      * @param array|null $cells Optional list of cells to be added to the newly created row.
      *